The First Commit of Many
-Adding JUL.py

Please refer to the README for information on how to modify this script.
This is a very primitive tool I built within a couple of hours used to log usernames that join a TikTok live stream. It serves as the foundation which many greater ideas & features will be built upon.